backend/requirements.txt
# AI Resume & Portfolio Builder — Backend Dependencies
# Python 3.13 compatible versions
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36

# ── Authentication ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic[email]==2.9.2

# ── AI / LLM ──────────────────────────────────────────────────────────────────
google-genai==1.5.0

# ── NLP / ATS Scoring ─────────────────────────────────────────────────────────
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.6.0
torchvision==0.21.0
numpy==2.1.3

# ── PDF Generation ─────────────────────────────────────────────────────────────
markdown2==2.5.0
xhtml2pdf==0.2.16
reportlab==4.2.5

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1

